requirements.txt
sqlalchemy==1.4.50
psycopg2-binary==2.9.9
langchain>=0.1.0
langgraph>=0.0.10
anthropic>=0.8.0
python-dotenv>=1.0.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-anthropic>=0.0.11
fastapi>=0.68.0
uvicorn>=0.15.0
websockets>=10.0
# Authentication dependencies
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
pydantic[email]>=2.0.0
# Email validation
email-validator>=2.0.0 